backend/requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic[email]==2.10.0

# CORS
python-multipart==0.0.19

# Authentication
PyJWT==2.10.1

# Database
supabase>=2.3.0
psycopg2-binary>=2.9.9
bcrypt>=4.1.2
redis>=5.0.1  # Redis caching

# Email
sendgrid==6.11.0

# ML/Experiment Tracking
wandb>=0.16.0  # Weights & Biases for AI tracking
# comet-ml>=3.48.0

# Error Monitoring
sentry-sdk[fastapi]>=1.40.0

# Testing
pytest>=7.4.3
pytest-cov>=4.1.0
pytest-asyncio>=0.21.1
httpx>=0.25.2  # For TestClient

# Original dependencies
blinker==1.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
Flask==3.1.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
requests==2.32.4
urllib3==2.5.0
Werkzeug==3.1.3
streamlit==1.47.1
stripe==12.4.0
python-dotenv==1.1.1
pandas==2.3.1
numpy==2.3.2
openai>=2.6.0
watchdog==6.0.0
